feat(server,types): send tracks inside create and join callback events
